Move search conftest to the root one
That's because we need search fixtures in other django apps as well.

This solves a bug with setup_es only disconnecting the seach
related signals at the end of the session causing some non-search
related tests to fail.
The signals are not connected and disconnected per test.
