--- datahub/cleanup/conftest.py
@@ -1,1 +1,0 @@
-from datahub.search.conftest import _es_client, _setup_es_indexes, setup_es  # noqa: F401
